chatbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
python-dotenv==1.0.0
psycopg[binary]==3.1.17
qdrant-client==1.7.3
tiktoken==0.5.2
langchain==0.1.5
langchain-openai==0.0.5
pydantic==2.5.3
httpx==0.26.0
